refactor: clean up quote checker

+++ mod.ts
@@ -4,7 +4,7 @@ import { config } from './config.ts';
 const client = new CommandClient({
 	prefix: '!',
 	owners: ['325254775828512778']
-});
+})
 
 client.on('ready', () => {
 	console.log('Ready!');
@@ -12,42 +12,19 @@ client.on('ready', () => {
 
 client.commands.loader.loadDirectory('./commands', { maxDepth: 2 });
 
-client.on('messageCreate', (message: Message) => {
+client.on('messageCreate', async (message: Message) => {
 	if (message.author.bot) return;
 
-	// Quote Checker
-	if (message.channelID == '789201783901650975') {
-		const fullYear = new Date(Date.now()).getFullYear();
-		const year = fullYear.toString().substring(2);
+	const year = new Date().getFullYear().toString().substring(2);
+	const QUOTE_REGEX = new RegExp(`"(.+)" - <@!?(\\d{17,19})> 2k${year}`);
 
-		if (!message.content.startsWith('"')) {
-			message.delete();
-			return message.channel.send('This message is not a quote.').then(async msg => {
-				await delay(2500);
-				msg.delete();
-			})
-		} else if (message.content.substring(1).startsWith('"')) {
-			message.delete();
-			return message.channel.send('Cannot use empty quote.').then(async msg => {
-				await delay(2500);
-				msg.delete();
-			})
-		} else if (!message.content.endsWith(year)) {
-			message.delete();
-			return message.channel.send(`This message does not end with 2k${year}.`).then(async msg => {
-				await delay(2500);
-				msg.delete();
-			})
-		} else if (!message.mentions.users.first()) {
-			message.delete();
-			return message.channel.send('This quote does not contain a mention.').then(async msg => {
-				await delay(2500);
-				msg.delete();
-			})
-		} else {
-			return;
-		}
+	// Quote Checker
+	if (message.channelID === '789201783901650975' && !QUOTE_REGEX.test(message.content)) {
+		message.delete();
+		const botMessage = await message.channel.send('This message is not a quote.');
+		await delay(2500);
+		botMessage.delete();
 	}
-})
+});
 
 client.connect(config.token, Intents.All);
